Make Extract useCallback() work in playground
There was an issue loading the parser and plugins to get ESLint to auto-fix
the code. The refactoring was failing because of that.

The trick was to override the ESLint methods to resolve the proper code.

It's a bit ugly, but it's the best we can do since we can't pass the parser
and plugins code directly to ESLint.