Fix Extract Variable scenario
It was messing up with the if() because the position of the declared variable
would be between the "if" keyword and the parenthesis.

I changed the position to always be at the beginning of the line.

Con: it won't be indented. But this is better than breaking the code.
Also, many codebases automatically reformat the code with Prettier nowadays.